refactor: extract shared syslog code into helpers/syslogutil
Move duplicated syslog writer construction, EnvelopeForMetric, and
retry logic into a shared package. Both metricsforwarder and
metricsgateway now delegate to syslogutil.NewSyslogWriter, reducing
code duplication to satisfy SonarCloud's 3% threshold.
